+++ source/engine/component/LightComponent.h
@@ -9,12 +9,12 @@ class LightComponent
 public:
 	LightComponent() :
 		m_lightType(lightType::POINT),
-		m_direction(vec4(0.0, 0.0, 0.0, 0.0)),
-		m_constantFactor(1.0),
-		m_linearFactor(0.14),
-		m_quadraticFactor(0.07),
-		m_radius(1.0),
-		m_color(vec4(1.0, 1.0, 1.0, 1.0)),
+		m_direction(vec4(0.0f, 0.0f, 0.0f, 0.0f)),
+		m_constantFactor(1.0f),
+		m_linearFactor(0.14f),
+		m_quadraticFactor(0.07f),
+		m_radius(1.0f),
+		m_color(vec4(1.0f, 1.0f, 1.0f, 1.0f)),
 		m_drawAABB(false) {};
 	~LightComponent() {};
 
@@ -24,10 +24,10 @@ public:
 	lightType m_lightType;
 	vec4 m_direction;
 
-	double m_constantFactor;
-	double m_linearFactor;
-	double m_quadraticFactor;
-	double m_radius;
+	float m_constantFactor;
+	float m_linearFactor;
+	float m_quadraticFactor;
+	float m_radius;
 
 	vec4 m_color;
